requirements.txt
# Core dependencies
tweepy>=4.12.0
python-dotenv>=0.21.0
openai>=1.0.0

# ML dependencies
torch>=2.0.0
transformers>=4.30.0

# MLOps dependencies
mlflow>=2.3.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Monitoring and logging
prometheus-client>=0.16.0
psutil>=5.9.0

# Utilities
requests>=2.28.0
pyyaml>=6.0
